requirements.txt
# FaceAna-MTQ 项目依赖包
# 包含核心功能和GUI界面开发所需的所有依赖
# 版本约束已优化以避免依赖冲突

# =============================================================================
# 核心GUI框架
# =============================================================================
flet>=0.21.0

# =============================================================================
# 图像处理
# =============================================================================
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.0,<2.0.0

# =============================================================================
# HTTP请求和API调用
# =============================================================================
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.9.0

# =============================================================================
# AI API SDKs
# =============================================================================
google-generativeai>=0.3.0
anthropic>=0.7.0
openai>=1.3.0

# =============================================================================
# 数据处理和可视化
# =============================================================================
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# =============================================================================
# 配置管理
# =============================================================================
python-dotenv>=1.0.0
pydantic>=2.5.0

# =============================================================================
# 异步处理
# =============================================================================
asyncio-mqtt>=0.13.0

# =============================================================================
# 系统监控和性能
# =============================================================================
psutil>=5.9.0

# =============================================================================
# 工具库
# =============================================================================
tqdm>=4.66.0
colorama>=0.4.6

# =============================================================================
# 开发和测试
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# =============================================================================
# 可选依赖 (根据需要取消注释)
# =============================================================================

# GPU加速支持 (需要CUDA环境)
# opencv-contrib-python>=4.8.0

# 高级图像处理
# scikit-image>=0.21.0

# 音频处理 (未来扩展)
# pyaudio>=0.2.11

# 视频处理 (未来扩展)
# moviepy>=1.0.3

# 机器学习 (未来扩展)
# scikit-learn>=1.3.0
# tensorflow>=2.13.0
# torch>=2.0.0